refactor(toolbox): use herbs toolbox for other commands
